+++ cycling_utils/cycling_utils/__init__.py
@@ -1,3 +1,4 @@
+from .saving import atomic_torch_save
 from .sampler import InterruptableDistributedSampler
 
-__all__ = ["InterruptableDistributedSampler"]
+__all__ = ["InterruptableDistributedSampler", "atomic_torch_save"]
